finished js for the challenge
I dont quite understand how to add and remove styles through js. for example, what i did was add a invalidate style but if you were to retype, the invalidate stays. I could work around it, by making it so that when you start typing it resets the whole page but that seems like a cheat. till next time!