finsihed base apparel styling
honestly, really displeased with this one. i had to "force" a lot of elements with translates and absolute positioning. So if it were scaled, it will look really bad. My issue was that the html needed to be in different places depending on whether it was mobile or desktop. and i couldn't figure out how to do it responsively so i just forced them individually.